// src/types/PortConflict.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PortRole } from "./PortRole";

export type PortConflict = { port: number, role: PortRole, pid: number, processName: string, executablePath?: string | null, instanceId?: string | null, instanceName?: string | null, };